fix up usages of controllers
